Move header extension validation into operation Validate method
