+++ p2panda-rs/src/operation/error.rs
@@ -16,6 +16,44 @@ pub enum OperationBuilderError {
     EncodeHeader(#[from] crate::operation::header::error::EncodeHeaderError),
 }
 
+#[derive(thiserror::Error, Debug)]
+pub enum ValidateHeaderExtensionsError {
+    /// Expect `timestamp` on all operations.
+    #[error("expected 'timestamp' in operation header")]
+    ExpectedTimestamp,
+
+    /// Expect `depth` on all operations.
+    #[error("expected 'depth' in operation header")]
+    ExpectedDepth,
+
+    /// Expected `previous` in UPDATE or DELETE operation.
+    #[error("expected 'previous' in UPDATE or DELETE operation")]
+    ExpectedPreviousOperations,
+
+    /// Unexpected `previous` in CREATE operation.
+    #[error("unexpected 'previous' in CREATE operation")]
+    UnexpectedPreviousOperations,
+
+    /// Expected `backlink` in UPDATE or DELETE operation.
+    #[error("expected 'backlink' in UPDATE or DELETE operation")]
+    ExpectedBacklink,
+
+    /// Unexpected `backlink` in CREATE operation.
+    #[error("unexpected 'backlink' in CREATE operation")]
+    UnexpectedBacklink,
+
+    /// Expected 'depth' to be 0 for CREATE operation.
+    #[error("expected 'depth' to be 0 for CREATE operation")]
+    ExpectedZeroDepth,
+
+    /// Expected 'depth' to be to be non-zero u64 for UPDATE and DELETE operations.
+    #[error("expected 'depth' to be non-zero u64 for UPDATE and DELETE operations")]
+    ExpectedNonZeroDepth,
+
+    #[error(transparent)]
+    HeaderValidation(#[from] crate::operation::header::error::ValidateHeaderError),
+}
+
 #[derive(Error, Debug)]
 pub enum ValidateOperationError {
     /// Expected `fields` in CREATE or UPDATE operation.
@@ -28,6 +66,9 @@ pub enum ValidateOperationError {
 
     #[error(transparent)]
     HeaderValidation(#[from] crate::operation::header::error::ValidateHeaderError),
+
+    #[error(transparent)]
+    HeaderExtensionValidation(#[from] crate::operation::error::ValidateHeaderExtensionsError),
 }
 
 /// Error types for methods of plain fields or operation fields.
